requirement.txt
django==1.11.0
django-filter==1.0.2
mysqlclient
